Feat get and post methods

+++ concessionaria/src/main/java/com/concessionaria/carro/CarroRepository.java
@@ -2,6 +2,6 @@ package com.concessionaria.carro;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarroRepository extends JpaRepository<Carro,Long>{
+public interface CarroRepository extends JpaRepository<Carro, Integer> {
 
 }
